Add first working recurrent example

+++ common/include/FloatingPointUtils.hpp
@@ -7,7 +7,7 @@
 namespace spacey::common {
 
 inline auto areFloatingPointEquals(const long double& first, const long double& second) -> bool {
-    return std::fabs(first - second) < std::numeric_limits<double>::epsilon();
+    return std::fabs(first - second) < std::numeric_limits<long double>::epsilon();
 }
 
 }
